% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFile.R
\name{LoadFile}
\alias{LoadFile}
\title{LoadFile function - road data multiful files in folder}
\usage{
LoadFile(x)
}
\arguments{
\item{x}{A text of folder path.}
}
\value{
The data of {x}
}
\description{
Load File function - road data multiful files in folder.
}
\author{
heeseon.han, https:/cafe.naver.com/dataan
}
\examples{
# "" : type the folder path
data<-LoadFile("")
}

